catch up

+++ daemon/openrazer_daemon/dbus_services/dbus_methods/chroma_keyboard.py
@@ -1,3 +1,5 @@
+# SPDX-License-Identifier: GPL-2.0-or-later
+
 """
 BlackWidow Chroma Effects
 """
@@ -15,14 +17,7 @@ def get_brightness(self):
     """
     self.logger.debug("DBus call get_brightness")
 
-    driver_path = self.get_driver_path('matrix_brightness')
-
-    with open(driver_path, 'r') as driver_file:
-        brightness = round(float(driver_file.read()) * (100.0 / 255.0), 2)
-
-        self.method_args['brightness'] = brightness
-
-        return brightness
+    return self.zone["backlight"]["brightness"]
 
 
 @endpoint('razer.device.lighting.brightness', 'setBrightness', in_sig='d')
@@ -39,11 +34,14 @@ def set_brightness(self, brightness):
 
     self.method_args['brightness'] = brightness
 
-    brightness = int(round(brightness * (255.0 / 100.0)))
-    if brightness > 255:
-        brightness = 255
+    if brightness > 100:
+        brightness = 100
     elif brightness < 0:
         brightness = 0
+
+    self.set_persistence("backlight", "brightness", int(brightness))
+
+    brightness = int(round(brightness * (255.0 / 100.0)))
 
     with open(driver_path, 'w') as driver_file:
         driver_file.write(str(brightness))
@@ -80,11 +78,14 @@ def set_game_mode(self, enable):
 
     driver_path = self.get_driver_path('game_led_state')
 
-    for kb_int in self.additional_interfaces:
-        super_file = os.path.join(kb_int, 'key_super')
-        alt_tab = os.path.join(kb_int, 'key_alt_tab')
-        alt_f4 = os.path.join(kb_int, 'key_alt_f4')
+    # Checks the usual MOUSE PROTOCOL Interface first
+    # before checking additional interfaces such as
+    # the KEYBOARD PROTOCOL.
+    super_file = self.get_driver_path('key_super')
+    alt_tab = self.get_driver_path('key_alt_tab')
+    alt_f4 = self.get_driver_path('key_alt_f4')
 
+    if os.path.exists(super_file):
         if enable:
             open(super_file, 'wb').write(b'\x01')
             open(alt_tab, 'wb').write(b'\x01')
@@ -93,6 +94,20 @@ def set_game_mode(self, enable):
             open(super_file, 'wb').write(b'\x00')
             open(alt_tab, 'wb').write(b'\x00')
             open(alt_f4, 'wb').write(b'\x00')
+    else:
+        for kb_int in self.additional_interfaces:
+            super_file = os.path.join(kb_int, 'key_super')
+            alt_tab = os.path.join(kb_int, 'key_alt_tab')
+            alt_f4 = os.path.join(kb_int, 'key_alt_f4')
+
+            if enable:
+                open(super_file, 'wb').write(b'\x01')
+                open(alt_tab, 'wb').write(b'\x01')
+                open(alt_f4, 'wb').write(b'\x01')
+            else:
+                open(super_file, 'wb').write(b'\x00')
+                open(alt_tab, 'wb').write(b'\x00')
+                open(alt_f4, 'wb').write(b'\x00')
 
     with open(driver_path, 'w') as driver_file:
         if enable:
@@ -128,6 +143,41 @@ def set_macro_mode(self, enable):
     self.logger.debug("DBus call set_macro_mode")
 
     driver_path = self.get_driver_path('macro_led_state')
+
+    with open(driver_path, 'w') as driver_file:
+        if enable:
+            driver_file.write('1')
+        else:
+            driver_file.write('0')
+
+
+@endpoint('razer.device.misc.keyswitchoptimization', 'getKeyswitchOptimization', out_sig='b')
+def get_keyswitch_optimization(self):
+    """
+    Get Keyswitch optimization state
+
+    :return: Status of keyswitch optimization
+    :rtype: bool
+    """
+    self.logger.debug("DBus call get_keyswitch_optimization")
+
+    driver_path = self.get_driver_path('keyswitch_optimization')
+
+    with open(driver_path, 'r') as driver_file:
+        return driver_file.read().strip() == '1'
+
+
+@endpoint('razer.device.misc.keyswitchoptimization', 'setKeyswitchOptimization', in_sig='b')
+def set_keyswitch_optimization(self, enable):
+    """
+    Set Keyswitch optimization state
+
+    :param enable: Status of keyswitch optimization
+    :type enable: bool
+    """
+    self.logger.debug("DBus call set_keyswitch_optimization")
+
+    driver_path = self.get_driver_path('keyswitch_optimization')
 
     with open(driver_path, 'w') as driver_file:
         if enable:
@@ -181,6 +231,10 @@ def set_wave_effect(self, direction):
     # Notify others
     self.send_effect_event('setWave', direction)
 
+    # remember effect
+    self.set_persistence("backlight", "effect", 'wave')
+    self.set_persistence("backlight", "wave_dir", int(direction))
+
     driver_path = self.get_driver_path('matrix_effect_wave')
 
     if direction not in self.WAVE_DIRS:
@@ -209,6 +263,10 @@ def set_static_effect(self, red, green, blue):
     # Notify others
     self.send_effect_event('setStatic', red, green, blue)
 
+    # remember effect
+    self.set_persistence("backlight", "effect", 'static')
+    self.zone["backlight"]["colors"][0:3] = int(red), int(green), int(blue)
+
     driver_path = self.get_driver_path('matrix_effect_static')
 
     payload = bytes([red, green, blue])
@@ -236,6 +294,10 @@ def set_blinking_effect(self, red, green, blue):
     # Notify others
     self.send_effect_event('setBlinking', red, green, blue)
 
+    # remember effect
+    self.set_persistence("backlight", "effect", 'blinking')
+    self.zone["backlight"]["colors"][0:3] = int(red), int(green), int(blue)
+
     driver_path = self.get_driver_path('matrix_effect_blinking')
 
     payload = bytes([red, green, blue])
@@ -254,6 +316,9 @@ def set_spectrum_effect(self):
     # Notify others
     self.send_effect_event('setSpectrum')
 
+    # remember effect
+    self.set_persistence("backlight", "effect", 'spectrum')
+
     driver_path = self.get_driver_path('matrix_effect_spectrum')
 
     with open(driver_path, 'w') as driver_file:
@@ -269,6 +334,9 @@ def set_none_effect(self):
 
     # Notify others
     self.send_effect_event('setNone')
+
+    # remember effect
+    self.set_persistence("backlight", "effect", 'none')
 
     driver_path = self.get_driver_path('matrix_effect_none')
 
@@ -316,8 +384,14 @@ def set_reactive_effect(self, red, green, blue, speed):
     # Notify others
     self.send_effect_event('setReactive', red, green, blue, speed)
 
+    # remember effect
+    self.set_persistence("backlight", "effect", 'reactive')
+    self.zone["backlight"]["colors"][0:3] = int(red), int(green), int(blue)
+
     if speed not in (1, 2, 3, 4):
         speed = 4
+
+    self.set_persistence("backlight", "speed", int(speed))
 
     payload = bytes([speed, red, green, blue])
 
@@ -334,6 +408,9 @@ def set_breath_random_effect(self):
 
     # Notify others
     self.send_effect_event('setBreathRandom')
+
+    # remember effect
+    self.set_persistence("backlight", "effect", 'breathRandom')
 
     driver_path = self.get_driver_path('matrix_effect_breath')
 
@@ -362,54 +439,13 @@ def set_breath_single_effect(self, red, green, blue):
     # Notify others
     self.send_effect_event('setBreathSingle', red, green, blue)
 
+    # remember effect
+    self.set_persistence("backlight", "effect", 'breathSingle')
+    self.zone["backlight"]["colors"][0:3] = int(red), int(green), int(blue)
+
     driver_path = self.get_driver_path('matrix_effect_breath')
 
     payload = bytes([red, green, blue])
-
-    with open(driver_path, 'wb') as driver_file:
-        driver_file.write(payload)
-
-
-@endpoint('razer.device.lighting.chroma', 'setBreathTriple', in_sig='yyyyyyyyy')
-def set_breath_triple_effect(self, red1, green1, blue1, red2, green2, blue2, red3, green3, blue3):
-    """
-    Set the device to dual colour breathing effect
-
-    :param red1: Red component
-    :type red1: int
-
-    :param green1: Green component
-    :type green1: int
-
-    :param blue1: Blue component
-    :type blue1: int
-
-    :param red2: Red component
-    :type red2: int
-
-    :param green2: Green component
-    :type green2: int
-
-    :param blue2: Blue component
-    :type blue2: int
-
-    :param red3: Red component
-    :type red3: int
-
-    :param green3: Green component
-    :type green3: int
-
-    :param blue3: Blue component
-    :type blue3: int
-    """
-    self.logger.debug("DBus call set_breath_dual_effect")
-
-    # Notify others
-    self.send_effect_event('setBreathDual', red1, green1, blue1, red2, green2, blue2, red3, green3, blue3)
-
-    driver_path = self.get_driver_path('matrix_effect_breath')
-
-    payload = bytes([red1, green1, blue1, red2, green2, blue2, red3, green3, blue3])
 
     with open(driver_path, 'wb') as driver_file:
         driver_file.write(payload)
@@ -443,9 +479,62 @@ def set_breath_dual_effect(self, red1, green1, blue1, red2, green2, blue2):
     # Notify others
     self.send_effect_event('setBreathDual', red1, green1, blue1, red2, green2, blue2)
 
+    # remember effect
+    self.set_persistence("backlight", "effect", 'breathDual')
+    self.zone["backlight"]["colors"][0:6] = int(red1), int(green1), int(blue1), int(red2), int(green2), int(blue2)
+
     driver_path = self.get_driver_path('matrix_effect_breath')
 
     payload = bytes([red1, green1, blue1, red2, green2, blue2])
+
+    with open(driver_path, 'wb') as driver_file:
+        driver_file.write(payload)
+
+
+@endpoint('razer.device.lighting.chroma', 'setBreathTriple', in_sig='yyyyyyyyy')
+def set_breath_triple_effect(self, red1, green1, blue1, red2, green2, blue2, red3, green3, blue3):
+    """
+    Set the device to triple colour breathing effect
+
+    :param red1: Red component
+    :type red1: int
+
+    :param green1: Green component
+    :type green1: int
+
+    :param blue1: Blue component
+    :type blue1: int
+
+    :param red2: Red component
+    :type red2: int
+
+    :param green2: Green component
+    :type green2: int
+
+    :param blue2: Blue component
+    :type blue2: int
+
+    :param red3: Red component
+    :type red3: int
+
+    :param green3: Green component
+    :type green3: int
+
+    :param blue3: Blue component
+    :type blue3: int
+    """
+    self.logger.debug("DBus call set_breath_triple_effect")
+
+    # Notify others
+    self.send_effect_event('setBreathTriple', red1, green1, blue1, red2, green2, blue2, red3, green3, blue3)
+
+    # remember effect
+    self.set_persistence("backlight", "effect", 'breathTriple')
+    self.zone["backlight"]["colors"][0:9] = int(red1), int(green1), int(blue1), int(red2), int(green2), int(blue2), int(red3), int(green3), int(blue3)
+
+    driver_path = self.get_driver_path('matrix_effect_breath')
+
+    payload = bytes([red1, green1, blue1, red2, green2, blue2, red3, green3, blue3])
 
     with open(driver_path, 'wb') as driver_file:
         driver_file.write(payload)
@@ -513,6 +602,10 @@ def set_ripple_effect(self, red, green, blue, refresh_rate):
     # Notify others
     self.send_effect_event('setRipple', red, green, blue, refresh_rate)
 
+    # remember effect
+    self.set_persistence("backlight", "effect", 'ripple')
+    self.zone["backlight"]["colors"][0:3] = int(red), int(green), int(blue)
+
 
 @endpoint('razer.device.lighting.custom', 'setRippleRandomColour', in_sig='d')
 def set_ripple_effect_random_colour(self, refresh_rate):
@@ -526,6 +619,9 @@ def set_ripple_effect_random_colour(self, refresh_rate):
 
     # Notify others
     self.send_effect_event('setRipple', None, None, None, refresh_rate)
+
+    # remember effect
+    self.set_persistence("backlight", "effect", 'rippleRandomColour')
 
 
 @endpoint('razer.device.lighting.chroma', 'setStarlightRandom', in_sig='y')
@@ -543,6 +639,10 @@ def set_starlight_random_effect(self, speed):
     # Notify others
     self.send_effect_event('setStarlightRandom')
 
+    # remember effect
+    self.set_persistence("backlight", "effect", 'starlightRandom')
+    self.set_persistence("backlight", "speed", int(speed))
+
 
 @endpoint('razer.device.lighting.chroma', 'setStarlightSingle', in_sig='yyyy')
 def set_starlight_single_effect(self, red, green, blue, speed):
@@ -559,6 +659,11 @@ def set_starlight_single_effect(self, red, green, blue, speed):
     # Notify others
     self.send_effect_event('setStarlightSingle', red, green, blue, speed)
 
+    # remember effect
+    self.set_persistence("backlight", "effect", 'starlightSingle')
+    self.set_persistence("backlight", "speed", int(speed))
+    self.zone["backlight"]["colors"][0:3] = int(red), int(green), int(blue)
+
 
 @endpoint('razer.device.lighting.chroma', 'setStarlightDual', in_sig='yyyyyyy')
 def set_starlight_dual_effect(self, red1, green1, blue1, red2, green2, blue2, speed):
@@ -574,3 +679,8 @@ def set_starlight_dual_effect(self, red1, green1, blue1, red2, green2, blue2, sp
 
     # Notify others
     self.send_effect_event('setStarlightDual', red1, green1, blue1, red2, green2, blue2, speed)
+
+    # remember effect
+    self.set_persistence("backlight", "effect", 'starlightDual')
+    self.set_persistence("backlight", "speed", int(speed))
+    self.zone["backlight"]["colors"][0:6] = int(red1), int(green1), int(blue1), int(red2), int(green2), int(blue2)
